pyproject.toml
[project]
name = "cpr-pivot-lab"
version = "0.1.0"
description = "CPR-ATR Strategy backtesting framework — DuckDB + Polars + AI Agent"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    # Agent Framework (Phidata + Ollama)
    "phidata>=2.7.10",
    "langchain-ollama>=1.0.1",
    "openai>=2.21.0",
    # Data layer: DuckDB for analytics, Polars for vectorized transforms
    "duckdb>=1.4.4",
    "polars>=0.20.0",
    "pyarrow>=18.0.0",
    # PostgreSQL: operational data only (sessions, signals, alerts)
    "psycopg[binary,pool]>=3.3.3",
    "sqlalchemy>=2.0.46",
    "anyio>=4.0.0",
    # Web / API
    "fastapi>=0.129.2",
    "uvicorn[standard]>=0.41.0",
    "jinja2>=3.1.0",
    "httpx>=0.28.0",
    "nicegui>=3.8.0",
    "plotly>=6.6.0",
    # Config & validation
    "pydantic>=2.12.5",
    "pydantic-settings>=2.13.1",
    "pyyaml>=6.0.2",
    "pytz>=2024.1",
    # Scheduling
    "apscheduler>=3.11.0",
    # Email
    "aiosmtplib>=3.0.0",
    "email-validator>=2.0.0",
    # Kite live data
    "kiteconnect>=5.0.0",
    "numpy>=2.2.6",
    # Progress tracking
    "tqdm>=4.66.0",
]

[project.scripts]
# Main AI agent CLI
pivot-agent = "agent.llm_agent:main"
# One-shot backtest runner
pivot-backtest = "engine.run_backtest:main"
# YAML-driven parameter sweep with auto-comparison
pivot-sweep = "engine.sweep_cli:main"
# CSV → Parquet converter (run once after downloading data)
pivot-convert = "engine.convert_to_parquet:main"
# Kite request_token exchange helper
pivot-kite-token = "scripts.kite_token:main"
# Alias for the same workflow, matching the command style used in nse-momentum-lab
pivot-kite-get-token = "scripts.kite_token:main"
# Kite historical ingestion into local parquet
pivot-kite-ingest = "scripts.kite_ingest:main"
# Build runtime tables
pivot-build = "scripts.build_tables:main"
# Initialize PostgreSQL schema
pivot-db-init = "scripts.init_postgres:main"
# Paper trading session control
pivot-paper-trading = "scripts.paper_trading:main"
# NiceGUI dashboard
pivot-dashboard = "web.run_nicedash:main"
# Data quality check
pivot-data-quality = "scripts.data_quality:main"
# Gold pipeline
pivot-gold = "scripts.gold_pipeline:main"
# Signal alerts
pivot-signal-alert = "scripts.signal_alert:main"
# Data validation
pivot-data-validate = "scripts.data_validate:main"
# Run parity checks between two saved run_ids
pivot-parity-check = "scripts.parity_check:main"
# Cleanup transient caches/logs/progress artifacts
pivot-clean = "scripts.clean_artifacts:main"
# Reset run history across DuckDB + PostgreSQL
pivot-reset-history = "scripts.reset_run_history:main"
# One-command refresh for build + optional paper prep
pivot-refresh = "scripts.refresh:main"
# Standardized long-window execution workflow (sequential strategy campaign)
pivot-campaign = "scripts.run_campaign:main"
# Data hygiene: detect and purge dead (delisted) symbols
pivot-hygiene = "scripts.data_hygiene:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["agent", "config", "db", "engine", "web", "scripts"]

[dependency-groups]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.24.0",
    "pytest-cov>=5.0.0",
"ruff>=0.8.0",
    "mypy>=1.12.0",
    "pre-commit>=4.5.1",
]

[tool.ruff]
line-length = 100
target-version = "py313"

[tool.ruff.lint]
select = ["E", "F", "W", "I", "N", "UP", "B", "C4", "RUF"]
ignore = ["E501"]

[tool.ruff.lint.per-file-ignores]
# web/components has extensive Plotly dict configs — C408 is acceptable there
"web/components/__init__.py" = ["C408"]
"web/pages/*.py" = ["C408"]  # Allow dict() for kpi_grid readability
# Web UI has user-facing text with ×, — –; these are intentional
"web/**/*.py" = ["RUF002", "RUF003"]
# Scripts and engine have user-facing console output and docs with ×, –
"scripts/*.py" = ["RUF002", "RUF003"]
"engine/*.py" = ["RUF002", "RUF003", "RUF010"]  # RUF010: Polars str() in f-strings
"agent/**/*.py" = ["RUF002", "RUF003"]
"db/*.py" = ["RUF002", "RUF003"]

[tool.mypy]
python_version = "3.14"
ignore_missing_imports = true
warn_return_any = false
# DuckDB fetchone() returns tuple|None; COUNT(*) always returns a row.
# Polars aggregation returns broad union types that are always numeric at runtime.
disable_error_code = ["index", "arg-type"]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
asyncio_mode = "auto"
cache_dir = ".tmp_pytest_cache"
addopts = "-p no:cacheprovider"
